Fix problems due to integer bit width or signedness
/home/richie/flexric/examples/xApp/c/orange/xapp_es_with_cell_util.c:1374:17: error: passing argument 1 of ‘forEachCell’ from incompatible pointer type [-Wincompatible-pointer-types]
 1374 |     forEachCell(getTargetCellID, doHandoverAction, switchOffCurrentCell, context);
      |                 ^~~~~~~~~~~~~~~
      |                 |
      |                 uint8_t (*)(callback_data_t) {aka unsigned char (*)(callback_data_t)}
/home/richie/flexric/examples/xApp/c/orange/xapp_es_with_cell_util.c:1155:27: note: expected ‘Callback’ {aka ‘short unsigned int (*)(callback_data_t)’} but argument is of type ‘uint8_t (*)(callback_data_t)’ {aka ‘unsigned char (*)(callback_data_t)’}
 1155 | void forEachCell(Callback targetCellFinding, Callback cbHOAction, Callback cbSwitchOffAction, callback_data_t data)
      |                  ~~~~~~~~~^~~~~~~~~~~~~~~~~

      warning: comparison of integer expressions of different signedness: ‘int’ and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]

+++ examples/xApp/c/orange/xapp_es_with_cell_util.c
@@ -280,12 +280,12 @@ void add_neighCell(struct SINRServingValues* UE, const uint16_t neighCellID, con
   }
 }
 
-uint8_t getTargetCellID(callback_data_t data) 
+uint16_t getTargetCellID(callback_data_t data)
 {
   assert(data.neighCells != NULL);
 
   double max_sinr = MIN_SINR;
-  uint8_t targetCell = 0;
+  uint16_t targetCell = 0;
 
   printf("\n=== Target Cell Selection for UE %d ===\n", data.ueID);
   printf("Current Cell: %d\n", data.frmCurntCell);
@@ -1176,7 +1176,7 @@ uint16_t doHandoverAction(callback_data_t data)
   return handover_sent ? 1 : 0;
 }
 
-uint8_t switchOffCurrentCell(callback_data_t data) 
+uint16_t switchOffCurrentCell(callback_data_t data)
 {
   rc_ctrl_req_data_t rc_ctrl = {0};
   ue_id_e2sm_t ue_id = gen_rc_ue_id(GNB_UE_ID_E2SM, data.ueID);
